added fragments and refactored MainActivity.
