+++ app/src/main/java/com/mornelytom/telegram/MainActivity.kt
@@ -2,32 +2,21 @@ package com.mornelytom.telegram
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.View
-import android.widget.Toast
 import androidx.appcompat.widget.Toolbar
-import com.mikepenz.materialdrawer.AccountHeader
-import com.mikepenz.materialdrawer.AccountHeaderBuilder
-import com.mikepenz.materialdrawer.Drawer
-import com.mikepenz.materialdrawer.DrawerBuilder
-import com.mikepenz.materialdrawer.model.DividerDrawerItem
-import com.mikepenz.materialdrawer.model.PrimaryDrawerItem
-import com.mikepenz.materialdrawer.model.ProfileDrawerItem
-import com.mikepenz.materialdrawer.model.interfaces.IDrawerItem
 import com.mornelytom.telegram.databinding.ActivityMainBinding
+import com.mornelytom.telegram.ui.fragments.ChatFragment
+import com.mornelytom.telegram.ui.objects.AppDrawer
 
 class MainActivity : AppCompatActivity() {
 
     private lateinit var mBinding: ActivityMainBinding
-    private lateinit var mDrawer: Drawer
-    private lateinit var mHeader: AccountHeader
+    private lateinit var mAppDrawer: AppDrawer
     private lateinit var mToolbar: Toolbar
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
         mBinding = ActivityMainBinding.inflate(layoutInflater)
-
         setContentView(mBinding.root)
     }
 
@@ -39,101 +28,16 @@ class MainActivity : AppCompatActivity() {
 
     private fun initFunc() {
         setSupportActionBar(mToolbar)
-        createHeader()
-        createDrawer()
-    }
-
-    private fun createHeader() {
-        mHeader = AccountHeaderBuilder()
-            .withActivity(this)
-            .withHeaderBackground(R.drawable.header)
-            .addProfiles(
-                ProfileDrawerItem()
-                    .withName("Senko San")
-                    .withEmail("+79000000000")
-            ).build()
-    }
-
-    private fun createDrawer() {
-        mDrawer = DrawerBuilder()
-            .withActivity(this)
-            .withToolbar(mToolbar)
-            .withActionBarDrawerToggle(true)
-            .withSelectedItem(-1)
-            .withAccountHeader(mHeader)
-            .addDrawerItems(
-                PrimaryDrawerItem()
-                    .withIdentifier(100)
-                    .withIconTintingEnabled(true)
-                    .withName("New Group")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_create_groups),
-                PrimaryDrawerItem()
-                    .withIdentifier(101)
-                    .withIconTintingEnabled(true)
-                    .withName("New Secret Chat")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_secret_chat),
-                PrimaryDrawerItem()
-                    .withIdentifier(102)
-                    .withIconTintingEnabled(true)
-                    .withName("New Channel")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_create_channel),
-                PrimaryDrawerItem()
-                    .withIdentifier(103)
-                    .withIconTintingEnabled(true)
-                    .withName("Contacts")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_contacts),
-                PrimaryDrawerItem()
-                    .withIdentifier(104)
-                    .withIconTintingEnabled(true)
-                    .withName("Calls")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_phone),
-                PrimaryDrawerItem()
-                    .withIdentifier(105)
-                    .withIconTintingEnabled(true)
-                    .withName("Saved Messages")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_favorites),
-                PrimaryDrawerItem()
-                    .withIdentifier(106)
-                    .withIconTintingEnabled(true)
-                    .withName("Settings")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_settings),
-                DividerDrawerItem(),
-                PrimaryDrawerItem()
-                    .withIdentifier(107)
-                    .withIconTintingEnabled(true)
-                    .withName("Invite Friends")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_invate),
-                PrimaryDrawerItem()
-                    .withIdentifier(108)
-                    .withIconTintingEnabled(true)
-                    .withName("Telegram FAQ")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_help)
+        mAppDrawer.create()
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.dataContainer,
+                ChatFragment()
             )
-            .withOnDrawerItemClickListener(object : Drawer.OnDrawerItemClickListener {
-                override fun onItemClick(
-                    view: View?,
-                    position: Int,
-                    drawerItem: IDrawerItem<*>
-                ): Boolean {
-                    Toast.makeText(applicationContext, position.toString(), Toast.LENGTH_SHORT).show()
-                    return false
-                }
-
-            })
-            .build()
+            .commit()
     }
 
     private fun initFields() {
         mToolbar = mBinding.mainToolbar
-
+        mAppDrawer = AppDrawer(this, mToolbar)
     }
 }
