+++ app/src/main/java/com/mornelytom/telegram/ui/fragments/BaseFragment.kt
@@ -1,0 +1,21 @@
+package com.mornelytom.telegram.ui.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+open class BaseFragment(val layout: Int) : Fragment() {
+
+    private lateinit var mRootView: View
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        mRootView = inflater.inflate(layout, container, false)
+        return mRootView
+    }
+
+}
