+++ app/src/main/java/com/mornelytom/telegram/ui/fragments/ChatFragment.kt
@@ -1,0 +1,10 @@
+package com.mornelytom.telegram.ui.fragments
+
+import com.mornelytom.telegram.R
+
+class ChatFragment : BaseFragment(R.layout.fragment_chat) {
+
+    override fun onResume() {
+        super.onResume()
+    }
+}
