+++ app/src/main/java/com/mornelytom/telegram/ui/fragments/SettingsFragment.kt
@@ -1,0 +1,10 @@
+package com.mornelytom.telegram.ui.fragments
+
+import com.mornelytom.telegram.R
+
+class SettingsFragment : BaseFragment(R.layout.fragment_settings)  {
+
+    override fun onResume() {
+        super.onResume()
+    }
+}
