+++ app/src/main/java/com/mornelytom/telegram/ui/objects/AppDrawer.kt
@@ -1,0 +1,123 @@
+package com.mornelytom.telegram.ui.objects
+
+import android.view.View
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
+import com.mikepenz.materialdrawer.AccountHeader
+import com.mikepenz.materialdrawer.AccountHeaderBuilder
+import com.mikepenz.materialdrawer.Drawer
+import com.mikepenz.materialdrawer.DrawerBuilder
+import com.mikepenz.materialdrawer.model.DividerDrawerItem
+import com.mikepenz.materialdrawer.model.PrimaryDrawerItem
+import com.mikepenz.materialdrawer.model.ProfileDrawerItem
+import com.mikepenz.materialdrawer.model.interfaces.IDrawerItem
+import com.mornelytom.telegram.R
+import com.mornelytom.telegram.ui.fragments.SettingsFragment
+
+class AppDrawer (val mainActivity: AppCompatActivity, val toolbar: Toolbar) {
+
+    private lateinit var mDrawer: Drawer
+    private lateinit var mHeader: AccountHeader
+
+    fun create() {
+        createHeader()
+        createDrawer()
+    }
+
+
+    private fun createHeader() {
+        mHeader = AccountHeaderBuilder()
+            .withActivity(mainActivity)
+            .withHeaderBackground(R.drawable.header)
+            .addProfiles(
+                ProfileDrawerItem()
+                    .withName("Senko San")
+                    .withEmail("+79000000000")
+            ).build()
+    }
+
+    private fun createDrawer() {
+        mDrawer = DrawerBuilder()
+            .withActivity(mainActivity)
+            .withToolbar(toolbar)
+            .withActionBarDrawerToggle(true)
+            .withSelectedItem(-1)
+            .withAccountHeader(mHeader)
+            .addDrawerItems(
+                PrimaryDrawerItem()
+                    .withIdentifier(100)
+                    .withIconTintingEnabled(true)
+                    .withName("New Group")
+                    .withSelectable(false)
+                    .withIcon(R.drawable.ic_menu_create_groups),
+                PrimaryDrawerItem()
+                    .withIdentifier(101)
+                    .withIconTintingEnabled(true)
+                    .withName("New Secret Chat")
+                    .withSelectable(false)
+                    .withIcon(R.drawable.ic_menu_secret_chat),
+                PrimaryDrawerItem()
+                    .withIdentifier(102)
+                    .withIconTintingEnabled(true)
+                    .withName("New Channel")
+                    .withSelectable(false)
+                    .withIcon(R.drawable.ic_menu_create_channel),
+                PrimaryDrawerItem()
+                    .withIdentifier(103)
+                    .withIconTintingEnabled(true)
+                    .withName("Contacts")
+                    .withSelectable(false)
+                    .withIcon(R.drawable.ic_menu_contacts),
+                PrimaryDrawerItem()
+                    .withIdentifier(104)
+                    .withIconTintingEnabled(true)
+                    .withName("Calls")
+                    .withSelectable(false)
+                    .withIcon(R.drawable.ic_menu_phone),
+                PrimaryDrawerItem()
+                    .withIdentifier(105)
+                    .withIconTintingEnabled(true)
+                    .withName("Saved Messages")
+                    .withSelectable(false)
+                    .withIcon(R.drawable.ic_menu_favorites),
+                PrimaryDrawerItem()
+                    .withIdentifier(106)
+                    .withIconTintingEnabled(true)
+                    .withName("Settings")
+                    .withSelectable(false)
+                    .withIcon(R.drawable.ic_menu_settings),
+                DividerDrawerItem(),
+                PrimaryDrawerItem()
+                    .withIdentifier(107)
+                    .withIconTintingEnabled(true)
+                    .withName("Invite Friends")
+                    .withSelectable(false)
+                    .withIcon(R.drawable.ic_menu_invate),
+                PrimaryDrawerItem()
+                    .withIdentifier(108)
+                    .withIconTintingEnabled(true)
+                    .withName("Telegram FAQ")
+                    .withSelectable(false)
+                    .withIcon(R.drawable.ic_menu_help)
+            )
+            .withOnDrawerItemClickListener(object : Drawer.OnDrawerItemClickListener {
+                override fun onItemClick(
+                    view: View?,
+                    position: Int,
+                    drawerItem: IDrawerItem<*>
+                ): Boolean {
+                    when (position) {
+                        7 -> mainActivity.supportFragmentManager.beginTransaction()
+                            .addToBackStack(null)
+                            .replace(R.id.dataContainer,
+                                SettingsFragment()
+                            )
+                            .commit()
+                    }
+                    return false
+                }
+
+            })
+            .build()
+    }
+}
